split up organization module
