+++ squarelet/organizations/models/__init__.py
@@ -1,0 +1,5 @@
+# pylint:disable=wildcard-import
+
+# Squarelet
+from squarelet.organizations.models.organization import *
+from squarelet.organizations.models.payment import *
